refactor(domain): 将比赛评分字段从Long类型改为Double类型
- 修改MatchPlayerStats实体类中的rating字段类型为Double
- 更新MatchPlayerStatsDO数据对象中的rating字段类型为Double
- 更新MatchPlayerStatsDTO传输对象中的rating字段类型为Double
- 相应修改构造函数参数类型
- 相应修改getter和setter方法返回值类型

+++ src/main/java/com/csxuhuan/gelatoni/application/dto/MatchPlayerStatsDTO.java
@@ -27,7 +27,7 @@ public class MatchPlayerStatsDTO {
     private String playerName;
 
     /** 比赛评分 */
-    private Long rating;
+    private Double rating;
 
     /** 是否MVP：1=是 */
     private Boolean isMvp;
@@ -117,11 +117,11 @@ public class MatchPlayerStatsDTO {
         this.playerName = playerName;
     }
 
-    public Long getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(Long rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
